% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cherryblossom-package.R
\docType{package}
\name{cherryblossom-package}
\alias{cherryblossom}
\alias{cherryblossom-package}
\title{cherryblossom: Cherry Blossom Run Race Results}
\description{
Race results of the Cherry Blossom Run, which is an annual road race that takes place in Washington, DC.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OpenIntroStat/cherryblossom}
  \item Report bugs at \url{https://github.com/OpenIntroStat/cherryblossom/issues}
}

}
\author{
\strong{Maintainer}: Mine Çetinkaya-Rundel \email{cetinkaya.mine@gmail.com} (\href{https://orcid.org/0000-0001-6452-2420}{ORCID})

}
\keyword{internal}
